+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/components/Navigation.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.mifos.core.common.utils.Constants
-import about.navigation.aboutScreen
+import com.mifos.feature.about.navigation.aboutScreen
 import com.mifos.feature.activate.navigation.activateScreen
 import com.mifos.feature.activate.navigation.navigateToActivateScreen
 import com.mifos.feature.auth.navigation.navigateToLogin
